backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
boto3
python-multipart
httpx
pytest
pytest-asyncio
python-dotenv
email-validator
